Fixed player rotation bug
Fixed the bug in which the player body would automatically begin to rotate and spin. This was because the Gun Mesh'x Box Collider was colliding with it and forcing it to rotate. The issue was resolved by de-activating the Box Collider's on the Gun Mesh and Gun Barrell, as I also do not want the player to die if bullets hit the gun.